Updated MapView
Added a more defined starting location on the map view. Currently auto
set to full sail university.

Played around with the region of the map, meaning how big of a area the
map will show when a zip code or the users location is set. Tried 500,
but that was too zoomed in, so I tried 1500, and it worked well.