Implemented A Parsing Method
Created a Parsing Method in the VC Extension file.

This Method takes the initial pull of a Zip code from the usda
database, then parses the son into a BroadMarket class and stores it
into an array of BroadMarket’s.

Next step is to parse the BroadMarket Information to achieve the
Address of each market so then it can be converted into Coordinates to
plot on the map.