Finalized API implementation
Fully added the USDA API’s implementation. The application will now
pull a specific amount of farmers markets that are associated with a
Zip code, parse that information from the API then re query the API and
get the markets deeper information from the market ID. To finally
convert the address to a CLLocationCoordinate2D and plot it on the map
with annotations.

Also added the Request for user’s location on launch of application.