Create Market Flow
Added the Whole Create market flow to the application.

Firebase is successfully storing user market creation.

Fixed some user flow bugs with google log in.

Added custom sliding controls to tableViewCell’s in the Callout View,
so a user can add them to their list, if they are logged in.