Fixed bugs, User search, User annotations
Fixed quite a few bugs with the overall flow of the application.

User search is fully implemented into the application.

Market creation has been in, but it was never plotting to the map as an
annotation. It is now plotting to the map if a user is in that area.